+++ IL2-SR-Common/Network/ReleaseMetadata.cs
@@ -7,7 +7,7 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Common.Network
         public const string GithubUserAgent = GithubUsername + "_" + GithubRepository;
 
         public const string Version = "1.0.4.7";
-        public const string ReleaseTag = "1.0.4.7-beta.2";
+        public const string ReleaseTag = "1.0.4.7-beta.3";
         public const string MinimumProtocolVersion = "1.0.0.0";
     }
 }