+++ IL2-SR-CommonTests/Input/InputBindingLayoutTests.cs
@@ -14,9 +14,10 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Common.Tests.Input
             Assert.IsTrue(InputBinding.Radio2ChannelUp >= InputBinding.Intercom);
             Assert.IsTrue(InputBinding.Radio2ChannelDown >= InputBinding.Intercom);
 
-            Assert.IsTrue(InputBinding.Radio1ChannelUp <= InputBinding.Radio2ChannelDown);
-            Assert.IsTrue(InputBinding.Radio1ChannelDown <= InputBinding.Radio2ChannelDown);
-            Assert.IsTrue(InputBinding.Radio2ChannelUp <= InputBinding.Radio2ChannelDown);
+            Assert.IsTrue(InputBinding.Radio1ChannelUp <= InputBinding.RestartSrs);
+            Assert.IsTrue(InputBinding.Radio1ChannelDown <= InputBinding.RestartSrs);
+            Assert.IsTrue(InputBinding.Radio2ChannelUp <= InputBinding.RestartSrs);
+            Assert.IsTrue(InputBinding.Radio2ChannelDown <= InputBinding.RestartSrs);
         }
 
         [TestMethod]
@@ -26,6 +27,19 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Common.Tests.Input
             Assert.AreEqual((int)InputBinding.Radio1ChannelDown + 100, (int)InputBinding.ModifierRadio1ChannelDown);
             Assert.AreEqual((int)InputBinding.Radio2ChannelUp + 100, (int)InputBinding.ModifierRadio2ChannelUp);
             Assert.AreEqual((int)InputBinding.Radio2ChannelDown + 100, (int)InputBinding.ModifierRadio2ChannelDown);
+        }
+
+        [TestMethod]
+        public void RestartSrsBindingIsInsideScannedInputRange()
+        {
+            Assert.IsTrue(InputBinding.RestartSrs >= InputBinding.Intercom);
+            Assert.IsTrue(InputBinding.RestartSrs >= InputBinding.Radio2ChannelDown);
+        }
+
+        [TestMethod]
+        public void RestartSrsModifierUsesStandardOffset()
+        {
+            Assert.AreEqual((int)InputBinding.RestartSrs + 100, (int)InputBinding.ModifierRestartSrs);
         }
     }
 }